refactor & restructure
