+++ app/src/main/AndroidManifest.xml
@@ -19,7 +19,7 @@
         android:theme="@style/Theme.Shotspots_v2"
         tools:targetApi="31" >
         <activity
-            android:name=".MainActivity"
+            android:name=".views.main.MainActivity"
             android:exported="true"
             android:label="@string/title_activity_main"
             android:theme="@style/Theme.Shotspots_v2" >
@@ -30,7 +30,7 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".views.locations.ImageUploadActivity"
+            android:name=".views.upload.ImageUploadActivity"
             android:exported="false" />
         <activity
             android:name=".views.map.MapActivity"
