+++ app/src/main/java/com/lbeul/shotspots_v2/views/main/LocationRecyclerViewAdapter.java
@@ -1,4 +1,4 @@
-package com.lbeul.shotspots_v2;
+package com.lbeul.shotspots_v2.views.main;
 
 import android.annotation.SuppressLint;
 import android.content.Context;
@@ -15,6 +15,7 @@ import android.widget.Toast;
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
 
+import com.lbeul.shotspots_v2.R;
 import com.lbeul.shotspots_v2.models.imageData.ImageData;
 import com.lbeul.shotspots_v2.models.inMemoryDatabase.DatabaseException;
 import com.lbeul.shotspots_v2.models.inMemoryDatabase.InMemoryDatabase;
